test(e2e): fix unstable server.port case
