+++ e2e/cases/server/dev.test.ts
@@ -85,49 +85,6 @@ rspackOnlyTest('default & hmr (default true)', async ({ page }) => {
   await rsbuild.close();
 });
 
-test('dev.port & output.distPath', async ({ page }) => {
-  const errors: string[] = [];
-  page.on('pageerror', (err) => errors.push(err.message));
-
-  const rsbuild = await dev({
-    cwd: join(fixtures, 'basic'),
-    plugins: [pluginReact()],
-    rsbuildConfig: {
-      dev: {
-        writeToDisk: true,
-      },
-      server: {
-        port: 3000,
-      },
-      output: {
-        distPath: {
-          root: 'dist-1',
-          js: 'aa/js',
-        },
-      },
-    },
-  });
-
-  await gotoPage(page, rsbuild);
-
-  expect(rsbuild.port).toBe(3000);
-
-  expect(
-    fse.existsSync(join(fixtures, 'basic/dist-1/index.html')),
-  ).toBeTruthy();
-
-  expect(fse.existsSync(join(fixtures, 'basic/dist-1/aa/js'))).toBeTruthy();
-
-  const locator = page.locator('#test');
-  await expect(locator).toHaveText('Hello Rsbuild!');
-
-  await rsbuild.close();
-
-  expect(errors).toEqual([]);
-
-  await fse.remove(join(fixtures, 'basic/dist-1'));
-});
-
 rspackOnlyTest(
   'hmr should work when setting dev.port & client',
   async ({ page }) => {
